Add Medium Java Solution for 875. Koko Eating Bananas from LeetCode

+++ LeetCode/Patterns/Binary Search/875. Koko Eating Bananas/Solution.java
@@ -1,6 +1,6 @@
 class Solution {
  public int bananas(int[] piles, int mid){
-    int hours = 0;{
+    long hours = 0;{
         for(int i =0;i<piles.length;i++){
             hours = hours + (piles[i] + mid -1)/mid;
         }
